升级权限架构---大部分完成，差部门管理全部接口和用户的部分接口修改

+++ helper-common/src/main/java/com/qc/printers/common/user/domain/entity/User.java
@@ -55,7 +55,7 @@ public class User implements Serializable {
     public Integer status;
 
     @ApiModelProperty(value = "部门")
-    public Integer deptId;
+    public Long deptId;
 //    public Long role;//权限更改为角色，再去查询角色所有的权限
 
     @ApiModelProperty(value = "电子邮箱")
